+++ lib/manual_speech_to_text.dart
@@ -1,4 +1,3 @@
 library manual_speech_to_text;
 
 export './src/manual_stt_controller.dart';
-export './src/manual_stt_service.dart';