+++ CHANGELOG.md
@@ -60,3 +60,12 @@
 ### Enhancements
 
 - Updated `README.md` content.
+
+## 1.0.0
+
+### Enhancements
+
+- Added `pauseIfMuteFor`, `clearTextOnStart` callbacks.
+- Optimised beef sound issue at some level.
+- And some other normal normal corrections.
+- Updated `README.md` content.